requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
python-dotenv
faiss-cpu
youtube-transcript-api
requests
sentence-transformers